Move proxy deps to proxy/
