+++ griptape_nodes_library/video/ltx_audio_to_video_generation.py
@@ -18,7 +18,7 @@ from griptape_nodes.exe_types.param_types.parameter_string import ParameterStrin
 from griptape_nodes.files.file import File, FileLoadError
 from griptape_nodes.traits.options import Options
 
-from griptape_nodes_library.griptape_proxy_node import GriptapeProxyNode
+from griptape_nodes_library.proxy import GriptapeProxyNode
 from griptape_nodes_library.utils.ffmpeg_utils import get_ffmpeg_path
 
 logger = logging.getLogger("griptape_nodes")
